BE_220930_myChallenge : 내참기 api 추가 완료

+++ backend/src/main/java/com/ssafy/kirin/dto/response/DonationDTO.java
@@ -1,4 +1,4 @@
 package com.ssafy.kirin.dto.response;
 
-public record DonationDTO(Integer amount, String nickname) {
+public record DonationDTO(Long amount, String nickname) {
 }
